change magic number for qmk hid host and vial conflict

+++ keyboards/ergohaven/hid.c
@@ -63,7 +63,7 @@ void raw_hid_receive_kb(uint8_t *data, uint8_t length) {
     bool res = process_raw_hid_data(data, length);
     if (res && is_keyboard_master()) transaction_rpc_send(RPC_SYNC_HID, length, data);
     if (res)
-        *((uint32_t*)data) = VIAL_HID_MAGIC;
+        *((uint64_t*)data) = VIAL_HID_MAGIC;
 }
 
 void hid_sync(uint8_t in_buflen, const void *in_data, uint8_t out_buflen, void *out_data) {
@@ -79,7 +79,7 @@ void keyboard_post_init_hid(void) {
 void raw_hid_receive_kb(uint8_t *data, uint8_t length) {
     bool res = process_raw_hid_data(data, length);
     if (res)
-        *((uint32_t*)data) = VIAL_HID_MAGIC;
+        *((uint64_t*)data) = VIAL_HID_MAGIC;
 }
 
 void keyboard_post_init_hid(void) {}
